Refactor PaymentMethod commands
