+++ src/Sylius/Bundle/ApiBundle/Command/GuestSChangePaymentMethod.php
@@ -11,11 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\Command\Checkout;
-
-use Sylius\Bundle\ApiBundle\Command\AbstractPaymentMethod;
+namespace Sylius\Bundle\ApiBundle\Command;
 
 /** @experimental */
-final class ChoosePaymentMethod extends AbstractPaymentMethod
+final class GuestSChangePaymentMethod extends AbstractPaymentMethod
 {
 }
